Separate to client and server.
It's possible to collect data in Raspberry or any other device and show
plots on another, if needed.

This commit implements client side only. Server implementation will
follow. Will bundle the client and server functionality into the same
binary so it's easy to distribute.

And while here, got an excuse to use protobuf and grpc for such purpose.

Signed-off-by: Ville Valkonen <weezel@users.noreply.github.com>